// src/app/api/swagger/models/api-report-type-enum.ts
/* tslint:disable */
/* eslint-disable */
export enum ApiReportTypeEnum {
  VipAnalyticsPurchase = 1,
  Supplier = 2,
  SuppliersList = 3,
  ImportPositions = 4,
  ParticipantOfferPublished = 5,
  TenantInformation = 6,
  CategoryTenant = 7,
  MarketTradesReport = 8,
  MarketSuppliersReport = 9,
  MarketCustomersReport = 10
}